Adjust build setup
