+++ package.json
@@ -3,6 +3,7 @@
   "scripts": {
     "prepare": "yarn workspace theme-ui prepare",
     "start": "yarn workspace docs start",
+    "clean": "yarn workspace docs clean",
     "test": "yarn workspace theme-ui test"
   },
   "workspaces": [
