+++ theme-ui/package.json
@@ -2,12 +2,12 @@
   "name": "theme-ui",
   "version": "0.0.11",
   "main": "dist/index.js",
-  "module": "dist/index.js",
+  "module": "dist/index.esm.js",
   "author": "Brent Jackson <jxnblk@gmail.com>",
   "license": "MIT",
   "repository": "system-ui/theme-ui",
   "scripts": {
-    "prepare": "microbundle",
+    "prepare": "microbundle --no-compress",
     "test": "jest"
   },
   "dependencies": {
@@ -17,6 +17,11 @@
     "lodash.get": "^4.4.2",
     "lodash.merge": "^4.6.1",
     "microbundle": "^0.11.0",
+    "@emotion/core": "^10.0.10",
+    "@emotion/styled": "^10.0.10",
+    "@mdx-js/react": "^1.0.0",
+    "emotion-theming": "^10.0.10",
+    "react": "^16.8.6",
     "styled-system": "^4.1.0",
     "typography": "^0.16.19"
   },
@@ -24,23 +29,11 @@
     "@babel/core": "^7.4.3",
     "@babel/preset-env": "^7.4.3",
     "@babel/preset-react": "^7.0.0",
-    "@emotion/core": "^10.0.10",
-    "@emotion/styled": "^10.0.10",
     "@mdx-js/mdx": "^1.0.0-rc.0",
-    "@mdx-js/react": "^1.0.0-rc.0",
-    "emotion-theming": "^10.0.10",
     "jest": "^24.7.0",
     "jest-emotion": "^10.0.10",
     "react-test-renderer": "^16.8.6",
     "typography-theme-wordpress-2016": "^0.16.19"
-  },
-  "peerDependencies": {
-    "@emotion/core": "^10.0.10",
-    "@emotion/styled": "^10.0.10",
-    "@mdx-js/mdx": "^1.0.0-rc.0",
-    "@mdx-js/react": "^1.0.0-rc.0",
-    "emotion-theming": "^10.0.10",
-    "react": "^16.8.6"
   },
   "jest": {
     "testRegex": "./test/.*.js$",
