+++ theme-ui/src/components.js
@@ -1,0 +1,54 @@
+import styled from '@emotion/styled'
+import { space } from 'styled-system'
+import themed from './themed'
+
+const tags = [
+  'p',
+  'a',
+  'h1',
+  'h2',
+  'h3',
+  'h4',
+  'h5',
+  'h6',
+  'img',
+  'pre',
+  'code',
+  'ol',
+  'ul',
+  'li',
+  'blockquote',
+  'hr',
+  'em',
+  'table',
+  'tr',
+  'th',
+  'td',
+  'em',
+  'strong',
+  'delete',
+  // mdx
+  'inlineCode',
+  'thematicBreak',
+  // other
+  'div',
+  // theme-ui
+  'root',
+]
+
+const aliases = {
+  inlineCode: 'code',
+  thematicBreak: 'hr',
+  root: 'div',
+}
+
+const alias = n => aliases[n] || n
+
+export const Styled = styled('div')(themed('div'), space)
+
+export const components = {}
+
+tags.forEach(tag => {
+  components[tag] = styled(alias(tag))(themed(tag), space)
+  Styled[tag] = components[tag]
+})
