+++ theme-ui/src/jsx.js
@@ -1,0 +1,14 @@
+import { jsx as emotion } from '@emotion/core'
+import css from '@styled-system/css'
+
+export const jsx = (type, props, ...children) =>
+  emotion.apply(undefined, [
+    type,
+    props ? ({
+      ...props,
+      css: css(props.css)
+    }) : null,
+    ...children
+  ])
+
+export default jsx
